LayoutTests/http/tests/contentextensions/main-resource-redirect-blocked-expected.txt

CONSOLE MESSAGE: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/resources/main-resource-redirect-blocked-target.html from loading a resource from http://127.0.0.1:8000/contentextensions/resources/main-resource-redirect-blocked-target.html
